Update tests for integer Decimal serialization
Update test expectations to match the new behavior where whole number
Decimal values are serialized as integers instead of floats. This aligns
with the change to _json_serializer that preserves integer types for
version numbers and other integer values.

+++ apollo/tests/test_admin_routes_supported_products.py
@@ -168,8 +168,8 @@ class TestJSONSerialization(unittest.TestCase):
         """Test JSON serializer with integer Decimal."""
         decimal_val = Decimal("42")
         result = _json_serializer(decimal_val)
-        self.assertEqual(result, 42.0)
-        self.assertIsInstance(result, float)
+        self.assertEqual(result, 42)
+        self.assertIsInstance(result, int)
 
     def test_json_serializer_unsupported_type(self):
         """Test JSON serializer with unsupported type."""
@@ -211,10 +211,11 @@ class TestJSONSerialization(unittest.TestCase):
 
         result = _format_export_data(data)
 
-        # Should be valid JSON with Decimals converted to floats
+        # Should be valid JSON with Decimals converted appropriately
+        # (floats for decimals, ints for whole numbers)
         parsed = json.loads(result)
         self.assertEqual(parsed[0]["price"], 19.99)
-        self.assertEqual(parsed[1]["price"], 99.0)
+        self.assertEqual(parsed[1]["price"], 99)
 
     def test_format_export_data_empty(self):
         """Test formatting empty export data."""
